requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
diffusers==0.27.2
dill==0.3.8
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.1
gradio==5.29.0
gradio_client==1.10.0
greenlet==3.1.0
groovy==0.1.2
grpcio==1.66.1
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.7
llvmlite==0.43.0
lpips==0.1.4
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.32.0
opencv-python==4.10.0.84
opensimplex==0.4.5.1
optuna==4.0.0
orjson==3.10.18
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==14.0.0
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
setuptools==72.1.0
shellingham==1.5.4
simplex==0.2.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.46.2
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.15
tomlkit==0.13.2
torch==2.4.1
torchmetrics==1.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.43.0
xxhash==3.5.0
yarl==1.11.0
zipp==3.20.1
